+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2026-06-23
+
+### Added
+
+- feat(nnmclub): anonymous-only support + credentials disclaimer
+
 ## [1.1.3] - 2026-06-22
 
 ### Fixed